fix(users): add create user test
This test was failing because the bcrypt library returns a different hash for the same password called repeatedly. Thus it is impossible to match the password entered by the user during tests. To solve this, write a custom matcher that implements the gomock matcher interface and use it to check params.